+++ src/logic/index.ts
@@ -1,2 +1,2 @@
-//We use this to generate types in Station
-export { generateTypesFile, generateTypingsForProfile } from './generate';
+//We use this in Air
+export * from './configure.utils';